Add a working solution
